postgres/requirements.txt
asyncpg==0.28.0
psycopg2-binary==2.9.9
